ai_document_assistant_backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
faiss-cpu
sentence-transformers
openai